# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: ParkXchange API
  version: 0.1.0
  description: |
    HTTP and WebSocket surface of the ParkXchange parking marketplace.
    This document is the source of truth for request and response types.
    Handlers stay handwritten on net/http; only these types are generated.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/reservations/{id}/driver-arrived:
    post:
      tags: [reservations]
      operationId: markDriverArrived
      summary: Record the driver's early “I arrived” signal
      parameters:
        - $ref: "#/components/parameters/ReservationID"
      responses:
        "204":
          description: Driver arrival recorded
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/reservations/{id}/clear-driver-arrived:
    post:
      tags: [reservations]
      operationId: clearDriverArrived
      summary: Retract the driver's “I arrived” signal so they can mark it again
      parameters:
        - $ref: "#/components/parameters/ReservationID"
      responses:
        "204":
          description: Driver arrival cleared
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /v1/reservations/{id}/driver-confirm-entered:
    post:
      tags: [reservations]
      operationId: driverConfirmEntered
      summary: Driver confirms entry after a stalled owner forgot “Salir ya”
      description: >
        Allowed after max(driver_ready_at, exchange_at) + 10 minutes. Completes
        the exchange and credits the owner.
      parameters:
        - $ref: "#/components/parameters/ReservationID"
      responses:
        "204":
          description: Exchange completed
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/reservations/{id}/driver-report-owner-no-show:
    post:
      tags: [reservations]
      operationId: driverReportOwnerNoShow
      summary: Driver reports the owner never left after the leave deadline
      description: >
        Allowed after max(driver_ready_at, exchange_at) + 10 minutes. Cancels
        with a full deposit release to the driver.
      parameters:
        - $ref: "#/components/parameters/ReservationID"
      responses:
        "204":
          description: Cancelled and deposit released
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ReservationID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Uniform error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
        request_id:
          type: string